+++ packages/data-schemas/src/admin/capabilities.ts
@@ -1,6 +1,10 @@
-import type { z } from 'zod';
-import type { TCustomConfig, configSchema } from 'librechat-data-provider';
-import { PrincipalType, PrincipalModel, ResourceType } from 'librechat-data-provider';
+import { ResourceType } from 'librechat-data-provider';
+import type {
+  BaseSystemCapability,
+  SystemCapability,
+  ConfigSection,
+  CapabilityCategory,
+} from '~/types/admin';
 
 // ---------------------------------------------------------------------------
 // System Capabilities
@@ -35,12 +39,6 @@ export const SystemCapabilities = {
   READ_ASSISTANTS: 'read:assistants',
   MANAGE_ASSISTANTS: 'manage:assistants',
 } as const;
-
-/** Base capabilities defined in the SystemCapabilities object. */
-export type BaseSystemCapability = (typeof SystemCapabilities)[keyof typeof SystemCapabilities];
-
-/** Principal types that can receive config overrides. */
-export type ConfigAssignTarget = 'user' | 'group' | 'role';
 
 /**
  * Capabilities that are implied by holding a broader capability.
@@ -109,28 +107,8 @@ export function expandImplications(directCaps: string[]): string[] {
 }
 
 // ---------------------------------------------------------------------------
-// Config section capabilities (derived from configSchema)
+// Resource & config capability mappings
 // ---------------------------------------------------------------------------
-
-/** Top-level keys of the configSchema from librechat.yaml. */
-export type ConfigSection = keyof z.infer<typeof configSchema>;
-
-/** Section-level config capabilities derived from configSchema keys. */
-type ConfigSectionCapability = `manage:configs:${ConfigSection}` | `read:configs:${ConfigSection}`;
-
-/** Principal-scoped config assignment capabilities. */
-type ConfigAssignCapability = `assign:configs:${ConfigAssignTarget}`;
-
-/**
- * Union of all valid capability strings:
- * - Base capabilities from SystemCapabilities
- * - Section-level config capabilities (manage:configs:<section>, read:configs:<section>)
- * - Config assignment capabilities (assign:configs:<user|group|role>)
- */
-export type SystemCapability =
-  | BaseSystemCapability
-  | ConfigSectionCapability
-  | ConfigAssignCapability;
 
 /**
  * Maps each ACL ResourceType to the SystemCapability that grants
@@ -166,98 +144,6 @@ export function configCapability(section: ConfigSection): `manage:configs:${Conf
 export function readConfigCapability(section: ConfigSection): `read:configs:${ConfigSection}` {
   return `read:configs:${section}`;
 }
-
-// ---------------------------------------------------------------------------
-// Admin API response types
-// ---------------------------------------------------------------------------
-
-/** Config document as returned by the admin API (no Mongoose internals). */
-export type AdminConfig = {
-  _id: string;
-  principalType: PrincipalType;
-  principalId: string;
-  principalModel: PrincipalModel;
-  priority: number;
-  overrides: Partial<TCustomConfig>;
-  isActive: boolean;
-  configVersion: number;
-  tenantId?: string;
-  createdAt?: string;
-  updatedAt?: string;
-};
-
-export type AdminConfigListResponse = {
-  configs: AdminConfig[];
-};
-
-export type AdminConfigResponse = {
-  config: AdminConfig;
-};
-
-export type AdminConfigDeleteResponse = {
-  success: boolean;
-};
-
-/** Audit action types for grant changes. */
-export type AuditAction = 'grant_assigned' | 'grant_removed';
-
-/** SystemGrant document as returned by the admin API. */
-export type AdminSystemGrant = {
-  id: string;
-  principalType: PrincipalType;
-  principalId: string;
-  capability: string;
-  grantedBy?: string;
-  grantedAt: string;
-  expiresAt?: string;
-};
-
-/** Audit log entry for grant changes as returned by the admin API. */
-export type AdminAuditLogEntry = {
-  id: string;
-  action: AuditAction;
-  actorId: string;
-  actorName: string;
-  targetPrincipalType: PrincipalType;
-  targetPrincipalId: string;
-  targetName: string;
-  capability: string;
-  timestamp: string;
-};
-
-/** Group as returned by the admin API. */
-export type AdminGroup = {
-  id: string;
-  name: string;
-  description: string;
-  memberCount: number;
-  topMembers: { name: string }[];
-  isActive: boolean;
-};
-
-/** Member entry as returned by the admin API for group/role membership lists. */
-export type AdminMember = {
-  userId: string;
-  name: string;
-  email: string;
-  avatarUrl?: string;
-  joinedAt: string;
-};
-
-/** Minimal user info returned by user search endpoints. */
-export type AdminUserSearchResult = {
-  userId: string;
-  name: string;
-  email: string;
-  avatarUrl?: string;
-};
-
-/** UI grouping of capabilities for the admin panel's capability editor. */
-export type CapabilityCategory = {
-  key: string;
-  labelKey: string;
-  capabilities: BaseSystemCapability[];
-};
 
 /** Pre-defined UI categories for grouping capabilities in the admin panel. */
 export const CAPABILITY_CATEGORIES: CapabilityCategory[] = [
