+++ packages/data-schemas/src/admin/index.ts
@@ -1,0 +1,1 @@
+export * from './capabilities';
