+++ packages/data-schemas/src/types/admin.ts
@@ -1,0 +1,138 @@
+import type { PrincipalType, PrincipalModel, TCustomConfig } from 'librechat-data-provider';
+import type { z, configSchema } from 'librechat-data-provider';
+
+/* ── Capability types ───────────────────────────────────────────────── */
+
+/** Base capabilities defined in the SystemCapabilities object. */
+export type BaseSystemCapability =
+  | 'access:admin'
+  | 'read:users'
+  | 'manage:users'
+  | 'read:groups'
+  | 'manage:groups'
+  | 'read:roles'
+  | 'manage:roles'
+  | 'read:configs'
+  | 'manage:configs'
+  | 'assign:configs'
+  | 'read:usage'
+  | 'read:agents'
+  | 'manage:agents'
+  | 'manage:mcpservers'
+  | 'read:prompts'
+  | 'manage:prompts'
+  | 'read:assistants'
+  | 'manage:assistants';
+
+/** Principal types that can receive config overrides. */
+export type ConfigAssignTarget = 'user' | 'group' | 'role';
+
+/** Top-level keys of the configSchema from librechat.yaml. */
+export type ConfigSection = keyof z.infer<typeof configSchema>;
+
+/** Section-level config capabilities derived from configSchema keys. */
+type ConfigSectionCapability = `manage:configs:${ConfigSection}` | `read:configs:${ConfigSection}`;
+
+/** Principal-scoped config assignment capabilities. */
+type ConfigAssignCapability = `assign:configs:${ConfigAssignTarget}`;
+
+/**
+ * Union of all valid capability strings:
+ * - Base capabilities from SystemCapabilities
+ * - Section-level config capabilities (manage:configs:<section>, read:configs:<section>)
+ * - Config assignment capabilities (assign:configs:<user|group|role>)
+ */
+export type SystemCapability =
+  | BaseSystemCapability
+  | ConfigSectionCapability
+  | ConfigAssignCapability;
+
+/** UI grouping of capabilities for the admin panel's capability editor. */
+export type CapabilityCategory = {
+  key: string;
+  labelKey: string;
+  capabilities: BaseSystemCapability[];
+};
+
+/* ── Admin API response types ───────────────────────────────────────── */
+
+/** Config document as returned by the admin API (no Mongoose internals). */
+export type AdminConfig = {
+  _id: string;
+  principalType: PrincipalType;
+  principalId: string;
+  principalModel: PrincipalModel;
+  priority: number;
+  overrides: Partial<TCustomConfig>;
+  isActive: boolean;
+  configVersion: number;
+  tenantId?: string;
+  createdAt?: string;
+  updatedAt?: string;
+};
+
+export type AdminConfigListResponse = {
+  configs: AdminConfig[];
+};
+
+export type AdminConfigResponse = {
+  config: AdminConfig;
+};
+
+export type AdminConfigDeleteResponse = {
+  success: boolean;
+};
+
+/** Audit action types for grant changes. */
+export type AuditAction = 'grant_assigned' | 'grant_removed';
+
+/** SystemGrant document as returned by the admin API. */
+export type AdminSystemGrant = {
+  id: string;
+  principalType: PrincipalType;
+  principalId: string;
+  capability: string;
+  grantedBy?: string;
+  grantedAt: string;
+  expiresAt?: string;
+};
+
+/** Audit log entry for grant changes as returned by the admin API. */
+export type AdminAuditLogEntry = {
+  id: string;
+  action: AuditAction;
+  actorId: string;
+  actorName: string;
+  targetPrincipalType: PrincipalType;
+  targetPrincipalId: string;
+  targetName: string;
+  capability: string;
+  timestamp: string;
+};
+
+/** Group as returned by the admin API. */
+export type AdminGroup = {
+  id: string;
+  name: string;
+  description: string;
+  memberCount: number;
+  topMembers: { name: string }[];
+  isActive: boolean;
+};
+
+/** Member entry as returned by the admin API for group/role membership lists. */
+export type AdminMember = {
+  userId: string;
+  name: string;
+  email: string;
+  avatarUrl?: string;
+  joinedAt: string;
+};
+
+/** Minimal user info returned by user search endpoints. */
+export type AdminUserSearchResult = {
+  userId: string;
+  name: string;
+  email: string;
+  avatarUrl?: string;
+};
