+++ packages/data-schemas/src/types/index.ts
@@ -30,6 +30,8 @@ export * from './systemGrant';
 export * from './group';
 /* Config */
 export * from './config';
+/* Admin */
+export * from './admin';
 /* Web */
 export * from './web';
 /* MCP Servers */
